refactor: organize into github directory
the future vision of this kernel is to be able to handle
webhook events from other platforms like telegram as well
