+++ package.json
@@ -11,7 +11,7 @@
   },
   "scripts": {
     "dev": "run-p worker proxy",
-		"predev": "lsof -i tcp:8787 | grep LISTEN | awk '{print $2}' | xargs kill -9",
+    "predev": "lsof -i tcp:8787 | grep LISTEN | awk '{print $2}' | xargs kill -9",
     "format": "run-s format:lint format:prettier format:cspell",
     "format:lint": "eslint --fix .",
     "format:prettier": "prettier --write .",
