+++ src/github/handlers/index.ts
@@ -1,6 +1,6 @@
-import { EventHandler } from "../event-handler";
+import { GitHubEventHandler } from "../github-event-handler";
 import { handleIssueCommentCreated } from "./issue/comment_created";
 
-export function bindHandlers(webhooks: EventHandler) {
+export function bindHandlers(webhooks: GitHubEventHandler) {
   webhooks.on("issue_comment.created", handleIssueCommentCreated);
 }
